+++ homematicip/aio/device.py
@@ -131,7 +131,9 @@ class AsyncShutterContact(ShutterContact, AsyncSabotageDevice):
     """ HMIP-SWDO (Door / Window Contact - optical) /
     HMIP-SWDO-I (Door / Window Contact Invisible - optical) """
 
-    pass
+
+class AsyncShutterContactOpticalPlus(ShutterContactOpticalPlus, AsyncShutterContact):
+    """ HmIP-SWDO-PL ( Window / Door Contact – optical, plus )"""
 
 
 class AsyncShutterContactMagnetic(ShutterContactMagnetic, AsyncDevice):
@@ -314,6 +316,8 @@ class AsyncFullFlushContactInterface(FullFlushContactInterface, AsyncDevice):
 
     pass
 
+class AsyncFullFlushInputSwitch(FullFlushInputSwitch, AsyncSwitch):
+    """ HMIP-FSI16 (Switch Actuator with Push-button Input 230V, 16A) """
 
 class AsyncFullFlushShutter(FullFlushShutter, AsyncDevice):
     """ HMIP-FROLL (Shutter Actuator - flush-mount) / HMIP-BROLL (Shutter Actuator - Brand-mount) """