+++ tests/test_devices.py
@@ -41,7 +41,7 @@ def test_room_control_device_analog(fake_home: Home):
     assert d.actualTemperature == 23.3
 
     assert str(d) == (
-        "ALPHA-IP-RBGa Raumbediengerät lowBat(False) unreach(False) rssiDeviceValue(-41) "
+        "ALPHA-IP-RBGa Raumbediengerät Analog lowBat(False) unreach(False) rssiDeviceValue(-41) "
         "rssiPeerValue(-29) configPending(False) dutyCycle(False) actualTemperature(23.3) "
         "setPointTemperature(23.0) temperatureOffset(0.0)"
     )
@@ -132,8 +132,28 @@ def test_full_flush_contact_interface(fake_home: Home):
     assert d.multiModeInputMode == MultiModeInputMode.KEY_BEHAVIOR
 
     assert str(d) == (
-        "HmIP-FCI1 Kontakt-Schnittstelle Unterputz – 1-fach lowBat(False) unreach(False) rssiDeviceValue(-46) rssiPeerValue(None) configPending(False) "
+        "HmIP-FCI1 Kontakt-Schnittstelle Unterputz – 1-fach lowBat(False) unreach(False) rssiDeviceValue(-46) "
+        "rssiPeerValue(None) configPending(False) "
         "dutyCycle(False) binaryBehaviorType(NORMALLY_CLOSE) multiModeInputMode(KEY_BEHAVIOR) windowState(CLOSED)"
+    )
+
+
+def test_full_flush_input_switch(fake_home: Home):
+    d = fake_home.search_device_by_id("3014F7110000000HmIPFSI16")
+    assert isinstance(d, FullFlushInputSwitch)
+
+    assert d.binaryBehaviorType == BinaryBehaviorType.NORMALLY_CLOSE
+    assert d.on == True
+    assert d.multiModeInputMode == MultiModeInputMode.KEY_BEHAVIOR
+    assert d.profileMode == ProfileMode.AUTOMATIC
+    assert d.userDesiredProfileMode == ProfileMode.AUTOMATIC
+
+    assert str(d) == (
+        "HmIP-FSI16 Wohnzimmer Beleuchtung lowBat(None) unreach(False) rssiDeviceValue(-57) "
+        "rssiPeerValue(-54) configPending(False) dutyCycle(False) deviceOverheated(False) "
+        "binaryBehaviorType(NORMALLY_CLOSE) multiModeInputMode(KEY_BEHAVIOR) on(True) "
+        "profileMode(AUTOMATIC) userDesiredProfileMode(AUTOMATIC) on(True) profileMode(AUTOMATIC) "
+        "userDesiredProfileMode(AUTOMATIC)"
     )
 
 
@@ -209,6 +229,23 @@ def test_shutter_device_magnetic(fake_home: Home):
     assert (
         d._rawJSONData
         == fake_home_download_configuration()["devices"]["3014F7110000000000005551"]
+    )
+
+
+def test_shutter_contact_optical_plus(fake_home: Home):
+    d = fake_home.search_device_by_id("3014F7110SHUTTER_OPTICAL")
+    assert isinstance(d, ShutterContactOpticalPlus)
+    assert d.label == "Sitzplatztüre"
+    assert d.windowState == WindowState.CLOSED
+    assert d.serializedGlobalTradeItemNumber == "3014F7110SHUTTER_OPTICAL"
+
+    assert str(d) == (
+        "HmIP-SWDO-PL Sitzplatztüre lowBat(False) unreach(False) rssiDeviceValue(-72) "
+        "rssiPeerValue(None) configPending(False) dutyCycle(False) sabotage(False) windowState(CLOSED)"
+    )
+    assert (
+        d._rawJSONData
+        == fake_home_download_configuration()["devices"]["3014F7110SHUTTER_OPTICAL"]
     )